Refactor preCheck routes into same route
