+++ packages/dito/app/cookies.server.ts
@@ -1,5 +1,5 @@
 import { createCookie } from "@remix-run/node";
-import type { Answers } from "routes/vorpruefung.$questionId";
+import type { Answers } from "routes/vorpruefung.($questionId)/route";
 
 export const userAnswers = createCookie("user-answers", {
   maxAge: 604_800, // one week
@@ -11,4 +11,8 @@ export const getAnswersFromCookie = async (request: Request) => {
     answers: Answers;
   } | null;
   return cookie ? cookie : { answers: {} };
+};
+
+export const getHeaderFromCookie = async (cookie: { answers: Answers }) => {
+  return { headers: { "Set-Cookie": await userAnswers.serialize(cookie) } };
 };
